+++ react/src/blueTheme.ts
@@ -8,7 +8,6 @@
 
 import { ThemeOptions } from '@material-ui/core/styles/createMuiTheme';
 import { typography, createSimpleLightPalette as createSimplePalette } from './shared';
-import shadows from './shadow';
 import * as PXBColors from '@pxblue/colors';
 import Color from 'color';
 
@@ -34,7 +33,7 @@ const ThemeColors = {
     text: {
         primary: PXBColors.black[500],
         secondary: PXBColors.gray[500],
-        // disabled: Color(PXBColors.black[300]).alpha(0.32).string(),
+        // disabled: Color(PXBColors.black[500]).alpha(0.3).string(),
         hint: PXBColors.gray[500],
     },
     action: {
@@ -67,22 +66,21 @@ export const blueTheme: ThemeOptions = {
         text: ThemeColors.text,
         action: ThemeColors.action,
     },
-    shadows: shadows,
     overrides: {
         // APP BAR OVERRIDES
         MuiAppBar: {
             colorDefault: {
                 color: ThemeColors.text.primary,
-                backgroundColor: PXBColors.white[50],
+                backgroundColor: ThemeColors.background.paper,
             },
             colorSecondary: {
-                color: PXBColors.white[50],
+                color: ThemeColors.background.paper,
                 backgroundColor: ThemeColors.primary.dark,
                 '& .MuiInputBase-root': {
-                    color: PXBColors.white[50],
+                    color: ThemeColors.background.paper,
                 },
                 '& .MuiSelect-icon': {
-                    color: PXBColors.white[50],
+                    color: ThemeColors.background.paper,
                 },
             },
         },
@@ -121,7 +119,7 @@ export const blueTheme: ThemeOptions = {
                 textTransform: 'none',
             },
             contained: {
-                backgroundColor: PXBColors.white[50],
+                backgroundColor: ThemeColors.background.paper,
                 color: ThemeColors.text.primary,
                 '&$disableElevation:not($containedPrimary):not($containedSecondary)': {
                     backgroundColor: PXBColors.white[500],
@@ -138,7 +136,7 @@ export const blueTheme: ThemeOptions = {
                 '&$disabled': {
                     backgroundColor: ThemeColors.background.paper,
                     border: `1px solid ${Color(PXBColors.black[500]).alpha(0.12).string()}`,
-                    color: Color(PXBColors.black[500]).alpha(0.3).string(),
+                    color: Color(ThemeColors.text.primary).alpha(0.3).string(),
                 },
             },
             containedPrimary: {
@@ -170,14 +168,14 @@ export const blueTheme: ThemeOptions = {
                 },
                 '&$disabled': {
                     backgroundColor: ThemeColors.background.paper,
-                    borderColor: Color(PXBColors.black[500]).alpha(0.12).string(),
-                    color: Color(PXBColors.black[500]).alpha(0.3).string(),
+                    borderColor: Color(BlackBorder).alpha(0.12).string(),
+                    color: Color(ThemeColors.text.primary).alpha(0.3).string(),
                 },
             },
             outlinedPrimary: {
                 borderColor: ThemeColors.primary.main,
                 '&$disabled': {
-                    borderColor: Color(PXBColors.black[500]).alpha(0.12).string(),
+                    borderColor: Color(BlackBorder).alpha(0.12).string(),
                 },
                 '&:hover': {
                     backgroundColor: Color(ThemeColors.primary.main).alpha(0.05).string(),
@@ -186,7 +184,7 @@ export const blueTheme: ThemeOptions = {
             outlinedSecondary: {
                 borderColor: ThemeColors.secondary.main,
                 '&$disabled': {
-                    borderColor: Color(PXBColors.black[500]).alpha(0.12).string(),
+                    borderColor: Color(BlackBorder).alpha(0.12).string(),
                 },
                 '&:hover': {
                     backgroundColor: Color(ThemeColors.secondary.main).alpha(0.05).string(),
@@ -263,7 +261,7 @@ export const blueTheme: ThemeOptions = {
                 },
                 '&$disabled': {
                     opacity: 1,
-                    color: Color(PXBColors.black[500]).alpha(0.3).string(),
+                    color: Color(ThemeColors.text.primary).alpha(0.3).string(),
                     '& $avatar': {
                         opacity: 0.5,
                     },
@@ -374,8 +372,8 @@ export const blueTheme: ThemeOptions = {
                 '&$disabled': {
                     opacity: 1,
                     backgroundColor: ThemeColors.background.paper,
-                    color: Color(PXBColors.black[500]).alpha(0.3).string(),
-                    borderColor: Color(PXBColors.black[500]).alpha(0.12).string(),
+                    color: ThemeColors.action.disabled,
+                    borderColor: Color(BlackBorder).alpha(0.12).string(),
                 },
             },
             outlinedSecondary: {
@@ -386,8 +384,8 @@ export const blueTheme: ThemeOptions = {
                 '&$disabled': {
                     opacity: 1,
                     backgroundColor: ThemeColors.background.paper,
-                    color: Color(PXBColors.black[500]).alpha(0.3).string(),
-                    borderColor: Color(PXBColors.black[500]).alpha(0.12).string(),
+                    color: ThemeColors.action.disabled,
+                    borderColor: Color(BlackBorder).alpha(0.12).string(),
                 },
             },
             icon: {
@@ -407,15 +405,15 @@ export const blueTheme: ThemeOptions = {
         MuiFab: {
             root: {
                 textTransform: 'none',
-                backgroundColor: PXBColors.white[50],
+                backgroundColor: ThemeColors.background.paper,
                 color: ThemeColors.text.primary,
                 '&:hover': {
                     backgroundColor: Color(PXBColors.black[500]).alpha(0.05).string(),
                 },
                 '&$disabled': {
                     backgroundColor: ThemeColors.background.paper,
-                    border: `1px solid ${Color(PXBColors.black[500]).alpha(0.12).string()}`,
-                    color: Color(PXBColors.black[500]).alpha(0.3).string(),
+                    border: `1px solid ${Color(BlackBorder).alpha(0.12).string()}`,
+                    color: ThemeColors.action.disabled,
                 },
             },
             primary: {
@@ -572,17 +570,48 @@ export const blueTheme: ThemeOptions = {
                 '&$checked + $track': {
                     opacity: 0.38,
                 },
+                '&$checked': {
+                    color: ThemeColors.secondary.main,
+                    '&$disabled': {
+                        color: Color(ThemeColors.secondary.main).mix(Color(ThemeColors.background.paper), 0.5).string(),
+                    },
+                    '&$disabled + $track': {
+                        backgroundColor: ThemeColors.secondary.main,
+                    },
+                },
             },
             colorPrimary: {
+                '&$disabled': {
+                    color: ThemeColors.background.paper,
+                },
+                '&$disabled + $track': {
+                    backgroundColor: PXBColors.black[100],
+                },
                 '&$checked': {
                     color: ThemeColors.primary.main,
+                    '&$disabled': {
+                        color: Color(ThemeColors.primary.main).mix(Color(ThemeColors.background.paper), 0.5).string(),
+                    },
+                    '&$disabled + $track': {
+                        opacity: 0.38,
+                        backgroundColor: Color(ThemeColors.primary.main)
+                            .mix(Color(ThemeColors.background.paper), 0.5)
+                            .string(),
+                    },
+                },
+            },
+            colorSecondary: {
+                '&$disabled': {
+                    color: ThemeColors.background.paper,
+                },
+                '&$disabled + $track': {
+                    backgroundColor: PXBColors.black[100],
                 },
             },
             track: {
                 backgroundColor: PXBColors.black[100],
-                opacity: 1,
+                opacity: 0.38,
             },
-            checked: {},
         },
 
         // TABLE OVERRIDES
@@ -598,14 +627,16 @@ export const blueTheme: ThemeOptions = {
         },
         MuiTableRow: {
             root: {
-                backgroundColor: '#FBFBFB', // possible new shade of white
+                backgroundColor: PXBColors.white[100],
                 '&$hover:hover': {
-                    backgroundColor: Color('#FBFBFB').mix(Color(PXBColors.black[50]), 0.5).string(),
+                    backgroundColor: Color(PXBColors.white[100]).mix(Color(PXBColors.black[50]), 0.5).string(),
                 },
                 '&:nth-of-type(odd):not($selected)': {
-                    backgroundColor: PXBColors.white[50],
+                    backgroundColor: ThemeColors.background.paper,
                     '&$hover:hover': {
-                        backgroundColor: Color(PXBColors.white[50]).mix(Color(PXBColors.black[50]), 0.5).string(),
+                        backgroundColor: Color(ThemeColors.background.paper)
+                            .mix(Color(PXBColors.black[50]), 0.5)
+                            .string(),
                     },
                 },
                 '&$selected': {
@@ -666,7 +697,7 @@ export const blueTheme: ThemeOptions = {
         },
         MuiTabs: {
             indicator: {
-                backgroundColor: PXBColors.white[50],
+                backgroundColor: ThemeColors.background.paper,
             },
         },
 
@@ -682,6 +713,9 @@ export const blueTheme: ThemeOptions = {
                     color: PXBColors.black[100],
                     opacity: 1,
                 },
+                '&:-webkit-autofill': {
+                    '-webkit-box-shadow': `0 0 0 30px ${PXBColors.white[50]} inset`,
+                },
             },
             adornedStart: {},
             adornedEnd: {},
@@ -694,7 +728,7 @@ export const blueTheme: ThemeOptions = {
                 },
                 '&:not($disabled):hover:before': {
                     borderBottomWidth: 1,
-                    borderBottomColor: PXBColors.black[500],
+                    borderBottomColor: BlackBorder,
                 },
                 '&$disabled:before': {
                     borderBottomColor: ThemeColors.divider,
@@ -724,6 +758,11 @@ export const blueTheme: ThemeOptions = {
                     pointerEvents: 'none',
                 },
             },
+            input: {
+                '&:-webkit-autofill': {
+                    '-webkit-box-shadow': `0 0 0 30px ${ThemeColors.background.default} inset`,
+                },
+            },
             underline: {
                 '&:before': {
                     borderBottomColor: ThemeColors.divider,
@@ -751,7 +790,7 @@ export const blueTheme: ThemeOptions = {
                     borderColor: ThemeColors.error.main,
                 },
                 '&$error:hover:not($focused) $notchedOutline': {
-                    borderColor: PXBColors.red[900],
+                    borderColor: ThemeColors.error.dark,
                 },
                 '&$disabled $notchedOutline': {
                     borderColor: ThemeColors.divider,