+++ angular/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v6.1.1 (June 30, 2021)
+
+### Fixed
+
+-   Fixed a few color issues in the `<pxb-drawer-header>` ([#184](https://github.com/pxblue/themes/issues/184)) and `<pxb-drawer-nav-item>` ([#186](https://github.com/pxblue/themes/issues/186)).
+
+### Changed
+
+-   Changed `mat-button` styles for disabled `primary`, `accent`, and `warn` variants.
+
 ## v6.1.0 (May 28, 2021)
 
 ### Added