+++ react/CHANGELOG.md
@@ -1,4 +1,12 @@
-# Change Log
+# Changelog
+
+## v6.1.0 (June 30, 2021)
+
+### Changed
+
+-   Update auto-fill styles for MUI `<TextInput>`.
+-   Update styles for disabled MUI `<Switch>`.
+-   Light theme shadow base color changed back to pure black (`#000000`).
 
 ## v6.0.0 (March 29, 2021)